#6: Move CityDataDTO to a separate class.
